used existing controller file and deleted newly created as per the pr comments
